Fix close stream, fix type value in Monitor to long.

+++ src/main/java/ru/innopolis/nikn/task1/IShareBoxMonitor.java
@@ -9,8 +9,7 @@ public interface IShareBoxMonitor {
      * Get current sum value
      * @return value int
      */
-    public int getValue();
-
+    public long getValue();
     /**
      * Add positive value to sum
      * @param value int
